Move status converter into spi
