+++ framework/bundles/org.jboss.tools.rsp.server/META-INF/MANIFEST.MF
@@ -33,6 +33,7 @@ Import-Package: org.eclipse.lsp4j.jsonrpc,
  org.jboss.tools.rsp.server.spi.model,
  org.jboss.tools.rsp.server.spi.model.polling,
  org.jboss.tools.rsp.server.spi.servertype,
+ org.jboss.tools.rsp.server.spi.util,
  org.osgi.framework,
  org.osgi.service.log,
  org.slf4j
