+++ runtimes/bundles/org.jboss.tools.rsp.server.wildfly/META-INF/MANIFEST.MF
@@ -34,6 +34,7 @@ Import-Package: org.eclipse.lsp4j.jsonrpc,
  org.jboss.tools.rsp.server.spi.model,
  org.jboss.tools.rsp.server.spi.model.polling,
  org.jboss.tools.rsp.server.spi.servertype,
+ org.jboss.tools.rsp.server.spi.util,
  org.osgi.framework,
  org.slf4j
 Export-Package: org.jboss.tools.rsp.server.wildfly.impl,
